Add InfoProperties
for Particle2DPanel and Particle3DPanel.
